# lib/api-zod/src/generated/types/createLeadRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateLeadRequestSource } from './createLeadRequestSource';

export interface CreateLeadRequest {
  /**
     * Phone number in any common format
     * @minLength 7
     * @maxLength 20
     */
  phone: string;
  /** Which form the lead came from */
  source?: CreateLeadRequestSource;
}